+++ src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './npmClient';
 export * from './packageManager';
 export * from './versionUtils';
 export * from './notify';
+export * from './logger';